SDC-12774. Amazon S3 origin exception when processing big files
Reviewed-on: https://review.streamsets.net/c/datacollector/+/28613
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ aws-lib/src/main/java/com/streamsets/pipeline/stage/origin/s3/AmazonS3SourceImpl.java
@@ -129,8 +129,8 @@ public class AmazonS3SourceImpl extends AbstractAmazonS3Source implements Amazon
 
       S3Offset offset = getOffsetFromGivenKey(s3Offset.getKey());
 
-      int offsetVal = AmazonS3Util.parseOffset(offset);
-      int s3offsetVal = AmazonS3Util.parseOffset(s3Offset);
+      long offsetVal = AmazonS3Util.parseOffset(offset);
+      long s3offsetVal = AmazonS3Util.parseOffset(s3Offset);
 
       if (!offset.getOffset().equals(S3Constants.MINUS_ONE) &&
           (s3Offset.getOffset().equals(S3Constants.MINUS_ONE) || s3offsetVal > offsetVal)) {
